# docs/swagger.yaml
definitions:
  handlers.ApiError:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  handlers.HeartbeatDTO:
    properties:
      is_alive:
        type: boolean
    type: object
  handlers.LoginDTO:
    properties:
      password:
        maxLength: 64
        minLength: 8
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  handlers.RegisterUserDTO:
    properties:
      password:
        maxLength: 64
        minLength: 8
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  handlers.SessionDocument:
    properties:
      _id:
        type: string
      apiKey:
        type: string
      eventId:
        type: string
      expiresAt:
        type: string
      userId:
        type: string
    type: object
  handlers.UserDocument:
    properties:
      _id:
        type: string
      createdAt:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: vorticalbox@protonmail.com
    name: Vorticalbox
  description: Bring Your Own Front End
  title: BYOFE
paths:
  /current-user:
    get:
      consumes:
      - application/json
      description: returns the currently logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserDocument'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ApiError'
      security:
      - WithToken: []
      summary: current user
      tags:
      - user
  /heartbeat:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HeartbeatDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ApiError'
      summary: Show the status of server.
      tags:
      - health
  /login:
    post:
      consumes:
      - application/json
      description: logs in a user and returns an API token
      parameters:
      - description: Login credentials
        in: body
        name: LoginDTO
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SessionDocument'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ApiError'
      summary: login
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: registers a new user
      parameters:
      - description: User registration details
        in: body
        name: RegisterUserDTO
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserDocument'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ApiError'
      summary: register user
      tags:
      - user
securityDefinitions:
  WithToken:
    description: token generated by the login endpoint
    in: header
    name: x-access-token
    type: apiKey
swagger: "2.0"